+++ package.json
@@ -27,6 +27,7 @@
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
     "react-switch": "^7.0.0",
+    "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.2",
     "typescript": "^4.9.5",
@@ -57,6 +58,7 @@
     ]
   },
   "devDependencies": {
-    "@types/mui-image": "^1.0.1"
+    "@types/mui-image": "^1.0.1",
+    "@types/redux-logger": "^3.0.9"
   }
 }
