+++ src/component/layout/common/TopBanner.tsx
